include/ntp_refclock.h: Change lencode to unsigned.
Lengths should not be able to go negative.  This shows refclock_gtlin()
should return unsigned.  Fixed a few odd bugs too.

+++ ntpd/ntp_control.c
@@ -1939,9 +1939,7 @@ ctl_putclock(
 		}
 		break;
 	case CC_TIMECODE:
-		ctl_putstr(CV_NAME,
-			   pcs->p_lastcode,
-			   (unsigned)pcs->lencode);
+		ctl_putstr(CV_NAME, pcs->p_lastcode, pcs->lencode);
 		break;
 
 	CASE_UINT(CC_POLL, pcs->polls);
